test(#1406): split oversized proposal test modules to meet F128 AC-X1
Why:
- Maintainer review flagged two additional PR-owned test modules that still
  violate the F128 AC-X1 hard limit of 350 lines:
  - packages/api/test/proposal-community-pr-gate.test.js: 478 lines
  - packages/api/test/proposal-stale-recovery.test.js: 382 lines

What changed:
- proposal-community-pr-gate.test.js (86 lines): kept only the no-PR-inference
  gate tests.
- proposal-source-envelope.test.js (new): source-envelope delivery, empty-string
  fallback, routing isolation, and Redis round-trip.
- proposal-seed-materialization.test.js (new): content-block preservation and
  missing-seed repair.
- proposal-stale-recovery.test.js (144 lines): kept stale-claim recovery tests.
- proposal-legacy-dedup.test.js (new): pre-idempotency legacy seed deduplication.
- proposal-reconcile-resilience.test.js (new): transient read-failure resilience.

Verification:
- pnpm --filter @cat-cafe/api build passes.
- node --test on all 11 affected proposal modules: 48 tests pass, 0 fail.
- biome check --write passes (only pre-existing cognitive complexity warnings in
  appendApprovedInitialMessage / executeQueuedDispatch remain).
