+++ packages/api/test/proposal-community-pr-gate.test.js
@@ -1,3 +1,12 @@
+// @ts-check
+/**
+ * F128 proposal runtime — no server-side PR inference.
+ *
+ * These tests verify that the proposal layer does not guess community-PR intent
+ * from title/reason text; the caller's explicit initialMessage and proposal
+ * fields are preserved verbatim.
+ */
+
 import assert from 'node:assert/strict';
 import { describe, test } from 'node:test';
 import './helpers/setup-cat-registry.js';
@@ -22,77 +31,6 @@ describe('F128 proposal runtime — no server-side PR inference', () => {
     const proposal = await ctx.proposalStore.get(proposalId);
     assert.equal(proposal.initialMessage, 'Own this PR and close the loop.');
     assert.equal(proposal.communityPrContext, undefined);
-  });
-
-  test('delivers a lossless source envelope to the child thread when initialMessage is omitted', async () => {
-    const ctx = await createProposalTestContext();
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-    const reason = 'Evaluate the existing community contribution.';
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: {
-        title: 'clowder-ai PR #1189 intake review',
-        reason,
-      },
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-    const proposal = await ctx.proposalStore.get(proposalId);
-    assert.equal(proposal.initialMessage, undefined);
-    assert.equal(proposal.communityPrContext, undefined);
-
-    const approveRes = await ctx.approve('alice', proposalId);
-    assert.equal(approveRes.statusCode, 200);
-    const { threadId } = JSON.parse(approveRes.body);
-
-    const timeline = await ctx.messageStore.getByThread(threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-    });
-    assert.equal(timeline.length, 1, 'child thread must receive a seed message even without initialMessage');
-    const seed = timeline[0];
-    assert.ok(seed.content.includes(reason), 'seed message must include the original reason');
-    assert.ok(
-      seed.content.includes('clowder-ai PR #1189 intake review'),
-      'seed message must include the original title',
-    );
-    assert.ok(seed.content.includes('## 主 Thread'), 'seed message must still include the fork-and-return header');
-    // The proposal record stores the raw (undefined) initialMessage, not the envelope.
-    assert.equal(proposal.initialMessage, undefined);
-  });
-
-  test('delivers reason-only PR URL to the child thread as a verifiable source envelope', async () => {
-    const ctx = await createProposalTestContext();
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-    const url = 'https://github.com/zts212653/clowder-ai/pull/1196';
-    const reason = `Advisory review of ${url} only.`;
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: {
-        title: 'Advisory discussion',
-        reason,
-      },
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-
-    const approveRes = await ctx.approve('alice', proposalId);
-    assert.equal(approveRes.statusCode, 200);
-    const { threadId } = JSON.parse(approveRes.body);
-
-    const timeline = await ctx.messageStore.getByThread(threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-    });
-    assert.equal(timeline.length, 1);
-    const seed = timeline[0];
-    assert.ok(seed.content.includes(url), 'child must be able to read the PR URL from the source envelope');
-    assert.ok(seed.content.includes('Advisory discussion'), 'seed message must include the original title');
-    assert.ok(seed.content.includes('## 主 Thread'), 'seed message must include the fork-and-return header');
   });
 
   test('does not treat repo-qualified shorthand as a PR without explicit review intent', async () => {
@@ -144,335 +82,5 @@ describe('F128 proposal runtime — no server-side PR inference', () => {
       assert.equal(proposal.initialMessage, undefined);
       assert.equal(proposal.communityPrContext, undefined);
     }
-  });
-
-  test('does not rewrite an unrelated internal proposal', async () => {
-    const ctx = await createProposalTestContext();
-    const source = await ctx.threadStore.create('alice', 'Internal work');
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: { initialMessage: 'Investigate the internal queue race.' },
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-    const proposal = await ctx.proposalStore.get(proposalId);
-    assert.equal(proposal.initialMessage, 'Investigate the internal queue race.');
-    assert.equal(proposal.communityPrContext, undefined);
-  });
-
-  test('treats empty-string initialMessage as absent and falls back to source envelope', async () => {
-    const ctx = await createProposalTestContext();
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-    const reason = 'Empty-string override must still deliver the source envelope.';
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: { title: 'Envelope fallback', reason, initialMessage: '' },
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-
-    const approveRes = await ctx.approve('alice', proposalId);
-    assert.equal(approveRes.statusCode, 200);
-    const { threadId } = JSON.parse(approveRes.body);
-
-    const timeline = await ctx.messageStore.getByThread(threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-    });
-    assert.equal(timeline.length, 1);
-    assert.ok(timeline[0].content.includes(reason), 'empty-string initialMessage must fall back to envelope content');
-  });
-
-  test('does not let @-mentions or #ideate inside title/reason influence routing', async () => {
-    const { InvocationQueue } = await import('../dist/domains/cats/services/agents/invocation/InvocationQueue.js');
-    const resolveCalls = [];
-    const router = {
-      async resolveTargetsAndIntent(content, threadId, options) {
-        resolveCalls.push({ content, threadId, options });
-        return { targetCats: [], intent: { intent: 'execute' }, hasMentions: false };
-      },
-    };
-    const ctx = await createProposalTestContext({
-      routerOverride: router,
-      invocationQueueOverride: new InvocationQueue(),
-      queueProcessorOverride: {
-        async processNext() {
-          return { started: true };
-        },
-      },
-    });
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: {
-        title: '@opus please review',
-        reason: '#ideate discuss this PR',
-        preferredCats: ['kimi'],
-      },
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-
-    const approveRes = await ctx.approve('alice', proposalId);
-    assert.equal(approveRes.statusCode, 200);
-
-    assert.equal(resolveCalls.length, 1, 'router must be consulted during approve dispatch');
-    assert.equal(
-      resolveCalls[0].content,
-      '',
-      'routing input must be empty when no explicit initialMessage is provided; title/reason must not leak',
-    );
-
-    const { threadId } = JSON.parse(approveRes.body);
-    const timeline = await ctx.messageStore.getByThread(threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-    });
-    assert.equal(timeline.length, 1);
-    const seed = timeline[0];
-    assert.ok(seed.content.includes('@opus'), 'seed content may contain the title text');
-    assert.ok(seed.content.includes('#ideate'), 'seed content may contain the reason text');
-    // The seed must still be serial (preferredCats[0] only) despite #ideate in reason.
-    assert.deepEqual(seed.mentions, ['kimi'], 'preferredCats[0] wins; #ideate in reason must not flip to parallel');
-  });
-
-  test('carries structured content blocks from the source message into the child seed', async () => {
-    const { InvocationQueue } = await import('../dist/domains/cats/services/agents/invocation/InvocationQueue.js');
-    const ctx = await createProposalTestContext({
-      invocationQueueOverride: new InvocationQueue(),
-      queueProcessorOverride: {
-        async processNext() {
-          return { started: true };
-        },
-      },
-    });
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-    const contentBlocks = [
-      {
-        type: 'file',
-        url: 'https://example.com/fix.patch',
-        fileName: 'fix.patch',
-        mimeType: 'text/x-diff',
-        fileSize: 123,
-      },
-    ];
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: { title: 'Patch review', reason: 'Please review the attached patch.' },
-      originContentBlocks: contentBlocks,
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-    const proposal = await ctx.proposalStore.get(proposalId);
-    assert.ok(proposal.sourceMessageId, 'proposal must record the source message id');
-
-    const approveRes = await ctx.approve('alice', proposalId);
-    assert.equal(approveRes.statusCode, 200);
-    const { threadId } = JSON.parse(approveRes.body);
-
-    const timeline = await ctx.messageStore.getByThread(threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-    });
-    assert.equal(timeline.length, 1);
-    const seed = timeline[0];
-    assert.ok(seed.content.includes('Patch review'), 'seed text must include title');
-    assert.deepEqual(seed.contentBlocks, contentBlocks, 'seed must preserve source message content blocks losslessly');
-    assert.ok(
-      seed.content.includes(proposal.sourceMessageId),
-      'seed text must expose the exact sourceMessageId so the child can dereference it',
-    );
-  });
-
-  test('preserves source content blocks even when an explicit initialMessage overrides the seed text', async () => {
-    const { InvocationQueue } = await import('../dist/domains/cats/services/agents/invocation/InvocationQueue.js');
-    const ctx = await createProposalTestContext({
-      invocationQueueOverride: new InvocationQueue(),
-      queueProcessorOverride: {
-        async processNext() {
-          return { started: true };
-        },
-      },
-    });
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-    const contentBlocks = [
-      {
-        type: 'file',
-        url: 'https://example.com/design.md',
-        fileName: 'design.md',
-        mimeType: 'text/markdown',
-        fileSize: 456,
-      },
-    ];
-    const initialMessage = 'Own this PR and review the attached design doc.';
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: {
-        title: 'External PR intake',
-        reason: 'Please review the attached design doc and decide on adoption.',
-        initialMessage,
-      },
-      originContentBlocks: contentBlocks,
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-    const proposal = await ctx.proposalStore.get(proposalId);
-    assert.ok(proposal.sourceMessageId, 'proposal must record the source message id');
-
-    const approveRes = await ctx.approve('alice', proposalId);
-    assert.equal(approveRes.statusCode, 200);
-    const { threadId } = JSON.parse(approveRes.body);
-
-    const timeline = await ctx.messageStore.getByThread(threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-    });
-    assert.equal(timeline.length, 1);
-    const seed = timeline[0];
-    assert.ok(seed.content.includes(initialMessage), 'seed must use the explicit initialMessage as the text body');
-    // #1387: explicit initialMessage must not hide the source envelope from the child prompt.
-    assert.ok(
-      seed.content.includes('External PR intake'),
-      'child prompt must still include the proposal title when explicit initialMessage is used',
-    );
-    assert.ok(
-      seed.content.includes('Please review the attached design doc and decide on adoption.'),
-      'child prompt must still include the proposal reason when explicit initialMessage is used',
-    );
-    assert.deepEqual(
-      seed.contentBlocks,
-      contentBlocks,
-      'explicit initialMessage must not discard source message content blocks',
-    );
-    assert.equal(
-      seed.extra?.crossPost?.sourceMessageId,
-      proposal.sourceMessageId,
-      'seed crossPost metadata must expose the exact sourceMessageId for child dereferencing',
-    );
-  });
-
-  test('source envelope and crossPost sourceMessageId survive a Redis round-trip', async () => {
-    const ctx = await createProposalTestContext();
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-    const url = 'https://github.com/zts212653/clowder-ai/pull/1197';
-    const initialMessage = 'Review this external PR.';
-
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: {
-        title: 'Redis round-trip intake',
-        reason: `Advisory review of ${url}.`,
-        initialMessage,
-      },
-    });
-
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-    const proposal = await ctx.proposalStore.get(proposalId);
-    assert.ok(proposal.sourceMessageId, 'proposal must record the source message id');
-
-    const approveRes = await ctx.approve('alice', proposalId);
-    assert.equal(approveRes.statusCode, 200);
-    const { threadId } = JSON.parse(approveRes.body);
-
-    // Re-fetch from Redis (not the in-memory cache) to verify round-trip persistence.
-    const timeline = await ctx.messageStore.getByThread(threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-    });
-    assert.equal(timeline.length, 1);
-    const seed = timeline[0];
-    assert.ok(seed.content.includes(initialMessage), 'round-tripped seed must include explicit initialMessage');
-    assert.ok(seed.content.includes(url), 'round-tripped seed must include the PR URL from the source envelope');
-    assert.ok(
-      seed.content.includes('Redis round-trip intake'),
-      'round-tripped seed must include the proposal title from the source envelope',
-    );
-    assert.equal(
-      seed.extra?.crossPost?.sourceMessageId,
-      proposal.sourceMessageId,
-      'round-tripped crossPost.sourceMessageId must match the proposal sourceMessageId',
-    );
-  });
-
-  test('repairs a missing child seed when the first dispatch fails after finalize, exactly once', async () => {
-    const ctx = await createProposalTestContext();
-    const source = await ctx.threadStore.create('alice', 'Community gatekeeper');
-    const res = await ctx.propose({
-      userId: 'alice',
-      threadId: source.id,
-      body: {
-        title: 'Repairable intake',
-        reason: 'Transient dispatch failure after finalize must be recoverable.',
-      },
-    });
-    assert.equal(res.statusCode, 200);
-    const { proposalId } = JSON.parse(res.body);
-
-    const originalAppend = ctx.messageStore.append.bind(ctx.messageStore);
-    let childAppendAttempts = 0;
-    ctx.messageStore.append = async function (input) {
-      // Fail only the child seed append, not source-thread messages.
-      if (input.threadId !== source.id) {
-        childAppendAttempts += 1;
-        if (childAppendAttempts === 1) {
-          throw new Error('simulated transient child-seed append failure');
-        }
-      }
-      return originalAppend(input);
-    };
-
-    const firstApprove = await ctx.approve('alice', proposalId);
-    assert.equal(firstApprove.statusCode, 200);
-    const firstBody = JSON.parse(firstApprove.body);
-    assert.ok(firstBody.threadId, 'thread must be created even when seed append fails');
-    assert.ok(
-      firstBody.warnings?.some((w) => w.includes('initialMessage append failed')),
-      'first approve must surface the seed failure as a warning',
-    );
-
-    let timeline = await ctx.messageStore.getByThread(firstBody.threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-      includeQueuedUserMessages: true,
-    });
-    assert.equal(timeline.length, 0, 'child seed must not exist after dispatch failure');
-
-    const secondApprove = await ctx.approve('alice', proposalId);
-    assert.equal(secondApprove.statusCode, 200);
-    const secondBody = JSON.parse(secondApprove.body);
-    assert.equal(secondBody.threadId, firstBody.threadId);
-    assert.equal(secondBody.deduped, true);
-
-    timeline = await ctx.messageStore.getByThread(firstBody.threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-      includeQueuedUserMessages: true,
-    });
-    assert.equal(timeline.length, 1, 'reconciled seed must exist exactly once');
-    assert.ok(
-      timeline[0].content.includes('Repairable intake'),
-      'reconciled seed must carry the source envelope title',
-    );
-
-    const thirdApprove = await ctx.approve('alice', proposalId);
-    assert.equal(thirdApprove.statusCode, 200);
-    const thirdBody = JSON.parse(thirdApprove.body);
-    assert.equal(thirdBody.deduped, true);
-
-    timeline = await ctx.messageStore.getByThread(firstBody.threadId, 10, 'alice', {
-      includeQueuedCatMessages: true,
-      includeQueuedUserMessages: true,
-    });
-    assert.equal(timeline.length, 1, 'further approves must not duplicate the seed');
   });
 });
